Code for assignment 2 finally finished

+++ Assignments/src/assignment1/Counter.java
@@ -6,6 +6,6 @@ public interface Counter {
 	
 	public void decrement();
 	
-	public int getCounter();
+	public long getCounter();
 
 }
